Deduplicate oauth functions
