+++ listenbrainz/domain/spotify.py
@@ -5,7 +5,9 @@ import time
 from flask import current_app
 import spotipy.oauth2
 
+from data.model.external_service import ExternalService
 from listenbrainz.db import spotify as db_spotify
+from listenbrainz.db import external_service_oauth as db_oauth
 from datetime import datetime, timezone
 
 SPOTIFY_API_RETRIES = 5
@@ -123,7 +125,9 @@ def refresh_user_token(spotify_user: Spotify):
     else:
         refresh_token = spotify_user.refresh_token
     expires_at = int(time.time()) + response['expires_in']
-    db_spotify.update_token(spotify_user.user_id, access_token, refresh_token, expires_at)
+    db_oauth.update_token(user_id=spotify_user.user_id, service=ExternalService.SPOTIFY,
+                          access_token=access_token, refresh_token=refresh_token,
+                          expires_at=expires_at)
     return get_user(spotify_user.user_id)
 
 
@@ -146,19 +150,20 @@ def get_user(user_id):
     Args:
         user_id (int): the ListenBrainz row ID of the user
     """
-    row = db_spotify.get_user(user_id)
+    row = db_oauth.get_token(user_id=user_id, service=ExternalService.SPOTIFY)
     if row:
         return Spotify.from_dbrow(row)
     return None
 
 
-def remove_user(user_id):
+def remove_user(user_id, stop_import=True):
     """ Delete user entry for user with specified ListenBrainz user ID.
 
     Args:
         user_id (int): the ListenBrainz row ID of the user
+        stop_import (bool): whether the (user, service) combination should be removed from the listens_importer table also
     """
-    db_spotify.delete_spotify(user_id)
+    db_oauth.delete_token(user_id=user_id, service=ExternalService.SPOTIFY, stop_import=stop_import)
 
 
 def add_new_user(user_id, spot_access_token):
@@ -172,10 +177,12 @@ def add_new_user(user_id, spot_access_token):
     access_token = spot_access_token['access_token']
     refresh_token = spot_access_token['refresh_token']
     expires_at = int(time.time()) + spot_access_token['expires_in']
-    permissions = spot_access_token['scope'].split()
-    active = SPOTIFY_IMPORT_PERMISSIONS[0] in permissions and SPOTIFY_IMPORT_PERMISSIONS[1] in permissions
+    scopes = spot_access_token['scope'].split()
+    active = SPOTIFY_IMPORT_PERMISSIONS[0] in scopes and SPOTIFY_IMPORT_PERMISSIONS[1] in scopes
 
-    db_spotify.create_spotify(user_id, access_token, refresh_token, expires_at, active, permissions)
+    db_oauth.save_token(user_id=user_id, service=ExternalService.SPOTIFY, access_token=access_token,
+                        refresh_token=refresh_token, token_expires_ts=expires_at,
+                        record_listens=active, scopes=scopes)
 
 
 def get_active_users_to_process():
@@ -294,8 +301,10 @@ class SpotifyInvalidGrantError(Exception):
 class SpotifyImporterException(Exception):
     pass
 
+
 class SpotifyListenBrainzError(Exception):
     pass
+
 
 class SpotifyAPIError(Exception):
     pass
